+++ u4-fonctions.h
@@ -76,12 +76,12 @@ void convergenceLigne(int j, pointeurFct fonction);
 //Prend en arg 2 unsigned long int et ressort le dégradé sur N dans un tableau de taille[N][3]
 void degradeRGB(unsigned long int A, unsigned long int  B,int N, int tab[][3]);
 void couleurs(unsigned long int A, unsigned long int B, unsigned long int C, int N1, int N2, int N3, unsigned long int tab[]);
-void couleursRGB(unsigned long int A, unsigned long int B, unsigned long int C, int N1, int N2, int N3, unsigned long int tab[][3]) ;
+void couleursRGB(unsigned long int A, unsigned long int B, unsigned long int C, int N1, int N2, int N3,int tab[][3]) ;
 
 //Enregistre une image de largeur Largeur au format PPM, dans un fichier Fichier en recalculant tout
 int enregistrerPPM(int Largeur, char Fichier[32]);
 
 void enregistrerParams(const char* fichier);
 void restaurerParams(const char* fichier);
-void calcBuffer(unsigned long int tabdeg[][3]);
+void calcBuffer(int tabdeg[][3]);
 #endif
